+++ Assignment 2 - part 6 - main game loop: fortification phase/FortificationPhase.cpp
@@ -7,3 +7,12 @@
 //
 
 #include "FortificationPhase.hpp"
+
+//Provide a group of C++ classes that implement the fortification phase following the official rules of the game of Risk.
+//In the fortification phase, the player is allowed to move a number of armies (X) from one of its countries (the source country) to one of its neighbors that it also owns (the target country).
+//X must be in the range [1 to (number of armies on the source country - 1)].
+//You must deliver a driver that demonstrates that
+//1) only valid countries can be selected as source/target;
+//2) only a valid number of armies can be moved;
+//3) the right number of armies is effectively moved.
+
